forgot to commit :(

+++ src/main/java/module1/Abilities/AttackAbility.java
@@ -2,7 +2,7 @@ package module1.Abilities;
 
 public class AttackAbility extends Ability {
 
-    public AttackAbility(int accuracy, float critChance, int potency, String element, String name) {
+    public AttackAbility(int accuracy, float critChance, float potency, String element, String name) {
         this.accuracy = accuracy;
         this.critChance = critChance;
         this.potency = potency;
